consolidate modules with similar functionality
